Add files via upload
Created a Flash Card App using Python Programming language that flips after 3 sec of pause to an English card after we have clicked the required button to a French flash card with change in background color .Using a file of words(French and English), fetched from a csv file through pandas and a GUI experience using Tk-inter. Additional feature: To keep track of words that are guessed wrong , to learn them in future.